[python] Make eos.data a real package
